Now body is executed in the current thread on LispWorks
This fixes issues of having signals from the body code signaled
in the incorrect thread.

Also, I've fixed usage of the LIFT:WHILE-MEASURING in tests.
From some moment it started to require a CATCH-ERROR-P argument.

Another error was fixed in the WHILE-MEASURING. It prevented it
from correct error catching. This pull-request should be merged
first:

https://github.com/gwkkwg/lift/pull/7